Fixing changes on webapps after merging all changes
