+++ sec/library/models/master.go
@@ -2,31 +2,27 @@ package models
 
 import (
 	"github.com/eaciit/orm"
-	"gopkg.in/mgo.v2/bson"
+	"sync"
 )
 
 // MasterUnitPlant
 type MasterUnitPlant struct {
-	orm.ModelBase `bson:"-" json:"-"`
-	Id            bson.ObjectId ` bson:"_id" json:"_id" `
-	Plant         string        `json:Plant`
-	Unit          string        `json:Unit`
+	sync.RWMutex
+	orm.ModelBase `bson:"-",json:"-"`
+	Plant         string `bson:"Plant" json:"Plant"`
+	Unit          string `bson:"Unit" json:"Unit"`
 }
 
 func NewMasterUnitPlant() *MasterUnitPlant {
 	m := new(MasterUnitPlant)
-	m.Id = bson.NewObjectId()
 	return m
-}
-
-func (e *MasterUnitPlant) RecordID() interface{} {
-	return e.Id
 }
 
 func (m *MasterUnitPlant) TableName() string {
 	return "MasterUnitPlant"
 }
 
+// MasterActivityType
 type MasterActivityType struct {
 	orm.ModelBase `bson:"-",json:"-"`
 	Id            string ` bson:"_id" , json:"_id" `
@@ -40,6 +36,7 @@ func (m *MasterActivityType) TableName() string {
 	return "MasterActivityType"
 }
 
+// MasterPhase
 type PhaseModel struct {
 	orm.ModelBase `bson:"-",json:"-"`
 	Id            int    ` bson:"Id" json:"Id" `
@@ -58,4 +55,15 @@ func (e *PhaseModel) RecordID() interface{} {
 
 func (m *PhaseModel) TableName() string {
 	return "MasterPhase"
+}
+
+// MasterUnit
+type MasterUnit struct {
+	sync.RWMutex
+	orm.ModelBase `bson:"-" json:"-"`
+	Unit          string `bson:"Unit" json:"Unit"`
+}
+
+func (m *MasterUnit) TableName() string {
+	return "MasterUnit"
 }
