--- sec/library/models/masteractivitytype.go
@@ -1,18 +1,0 @@
-package models
-
-import (
-	"github.com/eaciit/orm"
-)
-
-type MasterActivityType struct {
-	orm.ModelBase `bson:"-",json:"-"`
-	Id            string ` bson:"_id" , json:"Id" `
-}
-
-func (e *MasterActivityType) RecordID() interface{} {
-	return e.Id
-}
-
-func (m *MasterActivityType) TableName() string {
-	return "MasterActivityType"
-}
