+++ sec/library/models/powerplantcoordinates.go
@@ -20,13 +20,13 @@ type PowerPlantCoordinates struct {
 	FuelTypes_Diesel      bool    `bson:"FuelTypes_Diesel" json:"FuelTypes_Diesel"`
 	FuelTypes_Gas         bool    `bson:"FuelTypes_Gas" json:"FuelTypes_Gas"`
 	GasTurbineUnit        int     `bson:"GasTurbineUnit" json:"GasTurbineUnit"`
-	GasTurbineCapacity    int     `bson:"GasTurbineCapacity" json:"GasTurbineCapacity"`
+	GasTurbineCapacity    float64 `bson:"GasTurbineCapacity" json:"GasTurbineCapacity"`
 	SteamUnit             int     `bson:"SteamUnit" json:"SteamUnit"`
-	SteamCapacity         int     `bson:"SteamCapacity" json:"SteamCapacity"`
+	SteamCapacity         float64 `bson:"SteamCapacity" json:"SteamCapacity"`
 	DieselUnit            int     `bson:"DieselUnit" json:"DieselUnit"`
 	DieselCapacity        int     `bson:"DieselCapacity" json:"DieselCapacity"`
 	CombinedCycleUnit     int     `bson:"CombinedCycleUnit" json:"CombinedCycleUnit"`
-	CombinedCycleCapacity int     `bson:"CombinedCycleCapacity" json:"CombinedCycleCapacity"`
+	CombinedCycleCapacity float64 `bson:"CombinedCycleCapacity" json:"CombinedCycleCapacity"`
 	Longitude             float64 `bson:"Longitude" json:"Longitude"`
 	Latitude              float64 `bson:"Latitude" json:"Latitude"`
 }
