+++ sec/library/models/valueequationcomparison.go
@@ -40,12 +40,18 @@ func (m *ValueEquationComparison) SetPayLoad(k *knot.WebContext) error {
 
 func (m *ValueEquationComparison) GetData(ctx *orm.DataContext, k *knot.WebContext) (interface{}, error) {
 	m.SetPayLoad(k)
-	result, DataValue := tk.M{}, []tk.M{}
+	type DataValue struct {
+		Revenue         float64 `json:'Revenue'`
+		MaintenanceCost float64 `json:'MaintenanceCost'`
+		OperatingCost   float64 `json:'OperatingCost'`
+		NetGeneration   float64 `json:'NetGeneration'`
+	}
+	result, Result := tk.M{}, []DataValue{}
 	c := ctx.Connection
 	ve := ValueEquation{}
 	query := []*dbox.Filter{}
-	query = append(query, dbox.Gte("Period.Dates", m.StartPeriod))
-	query = append(query, dbox.Lte("Period.Dates", m.EndPeriod))
+	query = append(query, dbox.Gte("Dates", m.StartPeriod))
+	query = append(query, dbox.Lte("Dates", m.EndPeriod))
 	if m.SelectedPlant != nil && len(m.SelectedPlant) > 0 {
 		query = append(query, dbox.In("Plant", m.SelectedPlant))
 	}
@@ -58,24 +64,24 @@ func (m *ValueEquationComparison) GetData(ctx *orm.DataContext, k *knot.WebConte
 
 	csr, e := c.NewQuery().
 		Where(query...).
-		Aggr(dbox.AggrSum, "$Revenue", "Revenue").
-		Aggr(dbox.AggrSum, "$MaintenanceCost", "MaintenanceCost").
-		Aggr(dbox.AggrSum, "$OperatingCost", "OperatingCost").
-		Aggr(dbox.AggrSum, "$NetGeneration", "NetGeneration").
-		From(ve.TableName()).Group("").Cursor(nil)
-	e = csr.Fetch(&DataValue, 0, false)
+		Aggr(dbox.AggrSum, "Revenue", "Revenue").
+		Aggr(dbox.AggrSum, "MaintenanceCost", "MaintenanceCost").
+		Aggr(dbox.AggrSum, "OperatingCost", "OperatingCost").
+		Aggr(dbox.AggrSum, "NetGeneration", "NetGeneration").
+		From(ve.TableName()).Cursor(nil)
+	e = csr.Fetch(&Result, 0, false)
 	if e != nil {
 		return nil, e
 	}
 	csr.Close()
 	result.Set("Index", m.Index)
-	result.Set("DataValue", DataValue)
+	result.Set("DataValue", Result)
 	return result, nil
 }
 func (m *ValueEquationComparison) GetUnitList(ctx *orm.DataContext, k *knot.WebContext) (interface{}, error) {
 	m.SetPayLoad(k)
 	mup := MasterUnitPlant{}
-	result, UnitData := tk.M{}, []tk.M{}
+	result, UnitData := tk.M{}, []MasterUnitPlant{}
 	c := ctx.Connection
 	query := []*dbox.Filter{}
 	if m.SelectedPlant != nil && len(m.SelectedPlant) > 0 {
@@ -83,14 +89,11 @@ func (m *ValueEquationComparison) GetUnitList(ctx *orm.DataContext, k *knot.WebC
 	}
 
 	csr, e := c.NewQuery().
-		Where(query...).
+		Where(query...).Select("Unit").
 		From(mup.TableName()).Group("Unit").Cursor(nil)
 	e = csr.Fetch(&UnitData, 0, false)
 	if e != nil {
 		return nil, e
-	}
-	for _, i := range UnitData {
-		i.Set("_id", i.Get("_id").(tk.M).Get("Unit"))
 	}
 	csr.Close()
 	result.Set("Index", m.Index)
