--- sec/webapp/models/MasterActivityType.go
@@ -1,19 +1,0 @@
-package models
-
-/*import (
-	"github.com/eaciit/orm"
-)
-
-type MasterActivityType struct {
-	orm.ModelBase `bson:"-",json:"-"`
-	Id            string ` bson:"_id" , json:"_id" `
-}
-
-func (e *MasterActivityType) RecordID() interface{} {
-	return e.Id
-}
-
-func (m *MasterActivityType) TableName() string {
-	return "MasterActivityType"
-}
-*/
