--- sec/webapp/models/MasterFailureCode.go
@@ -1,20 +1,0 @@
-package models
-
-/*import (
-	"github.com/eaciit/orm"
-)
-
-type MasterFailureCode struct {
-	orm.ModelBase `bson:"-",json:"-"`
-	Id            string ` bson:"_id" , json:"_id" `
-	Text          string `bson:"text", json:"text"`
-}
-
-func (e *MasterFailureCode) RecordID() interface{} {
-	return e.Id
-}
-
-func (m *MasterFailureCode) TableName() string {
-	return "MasterFailureCode"
-}
-*/
