--- sec/webapp/models/MasterMROElement.go
@@ -1,21 +1,0 @@
-package models
-
-/*import (
-	"github.com/eaciit/orm"
-	"gopkg.in/mgo.v2/bson"
-)
-
-type MasterMROElement struct {
-	orm.ModelBase `bson:"-",json:"-"`
-	Id            bson.ObjectId ` bson:"_id" , json:"_id" `
-	Element       string        `bson:"Element", json:"Element"`
-}
-
-func (e *MasterMROElement) RecordID() interface{} {
-	return e.Id
-}
-
-func (m *MasterMROElement) TableName() string {
-	return "MasterMROElement"
-}
-*/
