--- sec/webapp/models/MasterOrderType.go
@@ -1,21 +1,0 @@
-package models
-
-/*import (
-	"github.com/eaciit/orm"
-	"gopkg.in/mgo.v2/bson"
-)
-
-type MasterOrderType struct {
-	orm.ModelBase `bson:"-",json:"-"`
-	Id            bson.ObjectId ` bson:"_id" , json:"_id" `
-	OrderTypeDesc string
-}
-
-func (e *MasterOrderType) RecordID() interface{} {
-	return e.Id
-}
-
-func (m *MasterOrderType) TableName() string {
-	return "MasterOrderType"
-}
-*/
