--- sec/webapp/models/phase.go
@@ -1,26 +1,0 @@
-package models
-
-import (
-	"github.com/eaciit/orm"
-	// "gopkg.in/mgo.v2/bson"
-)
-
-type PhaseModel struct {
-	orm.ModelBase `bson:"-",json:"-"`
-	Id            int    ` bson:"Id" json:"Id" `
-	Phase         string ` bson:"Phase" json:"Phase" `
-}
-
-func NewPhaseModel() *PhaseModel {
-	m := new(PhaseModel)
-	// m.Id = bson.NewObjectId()
-	return m
-}
-
-func (e *PhaseModel) RecordID() interface{} {
-	return e.Id
-}
-
-func (m *PhaseModel) TableName() string {
-	return "MasterPhase"
-}
