+++ sec/webapp/views/valueequation/historicalvalueequation/dataquality.html
@@ -142,6 +142,7 @@ HVEDataQuality.GetData = function(){
 				var DataQuality = res.Data;
 				if(HVESummary.scope()=="Plant"){
 					if(ValueEquation.selected().length==1){
+						console.log(DataQuality)
 						for(var i in DataQuality){
 							DataQuality[i].ID = DataQuality[i].Unit;
 							DataQuality[i].Count = 1;
@@ -175,7 +176,7 @@ HVEDataQuality.GetData = function(){
 				}
 				if (ValueEquation.scope()!=='Unit'){
 					HVESummary.BreakdownBy('Plant');
-					// if(ValueEquation.selected().length!==1){
+					if(ValueEquation.scope() !== 'Plant'){
 						for(var i in DataQuality){
 							DataQuality[i].CapacityPaymentDocuments = [];
 							DataQuality[i].EnergyPaymentDocuments = [];
@@ -188,24 +189,13 @@ HVEDataQuality.GetData = function(){
 							DataQuality[i].PrimaryFuel2ndDocuments = [];
 							DataQuality[i].BackupFuelDocuments = [];
 						}
-					// }
+					}
 				}else{
 					HVESummary.BreakdownBy('Unit');
 					for(var i in DataQuality){
 						DataQuality[i].ID = DataQuality[i].Unit;
 						DataQuality[i].Count = 1;
 
-
-						// DataQuality[i].CapacityPaymentDocuments = [];
-						// DataQuality[i].EnergyPaymentDocuments = [];
-						// DataQuality[i].StartupPaymentDocuments = [];
-						// DataQuality[i].PenaltyDocuments = [];
-						// DataQuality[i].IncentiveDocuments = [];
-						// DataQuality[i].MaintenanceCostDocuments = [];
-						// DataQuality[i].MaintenanceDurationDocuments = [];
-						// DataQuality[i].PrimaryFuel1stDocuments = [];
-						// DataQuality[i].PrimaryFuel2ndDocuments = [];
-						// DataQuality[i].BackupFuelDocuments = [];
 					}
 				}
 				HVEDataQuality.DataQuality(DataQuality);
